+++ src/main/java/org/jolokia/docker/maven/StartMojo.java
@@ -102,7 +102,7 @@ public class StartMojo extends AbstractDockerMojo {
                     .memorySwap(runConfig.getMemorySwap())
                     .entrypoint(runConfig.getEntrypoint())
                     .exposedPorts(mappedPorts.getContainerPorts())
-                    .environment(runConfig.getEnv())
+                    .environment(runConfig.getEnvPropertyFile(),runConfig.getEnv())
                     .command(runConfig.getCommand())
                     .hostConfig(createContainerHostConfig(docker, runConfig, mappedPorts));
             VolumeConfiguration volumeConfig = runConfig.getVolumeConfiguration();