+++ src/main/java/org/jolokia/docker/maven/config/handler/property/ConfigKey.java
@@ -40,6 +40,7 @@ public enum ConfigKey {
     DNS_SEARCH,
     ENTRYPOINT,
     ENV,
+    ENV_PROPERTY_FILE,
     EXTRA_HOSTS,
     FROM,
     HOSTNAME,